add H2 database

+++ src/main/java/com/aktivgo/demo/dao/Dao.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface Dao<T> {
-    int size();
+    long size();
 
-    Optional<T> get(int id);
+    Optional<T> get(long id);
 
     List<T> getAll();
